Allow engines to live in lib/. Clean up, remove unneeded complexity.
